[serve] Separate autoscaling unit tests (#42301)
